+++ examples/server.rs
@@ -21,7 +21,7 @@ async fn server(
 
 #[ntex::main]
 async fn main() -> std::io::Result<()> {
-    std::env::set_var("RUST_LOG", "ntex=trace,ntex_amqp=trace,basic=trace");
+    std::env::set_var("RUST_LOG", "trace,ntex_io=info");
     env_logger::init();
 
     ntex::server::Server::build()